--- a/opentelemetry-sdk/src/metrics/internal/aggregate.rs
+++ b/opentelemetry-sdk/src/metrics/internal/aggregate.rs
@@ -9,9 +9,13 @@ use crate::{
 };
 
 use super::{
+    exponential_histogram,
     exponential_histogram::ExpoHistogram,
+    histogram,
     histogram::Histogram,
+    last_value,
     last_value::LastValue,
+    sum,
     sum::{PrecomputedSum, Sum},
     Number,
 };
@@ -38,6 +42,34 @@ where
 {
     fn call(&self, measurement: T, attrs: AttributeSet) {
         self(measurement, attrs)
+    }
+}
+
+/// Allows creating a bounded measure based on the specified attribute set
+pub(crate) trait BoundedMeasureGenerator<T>: Send + Sync + 'static {
+    fn generate(&self, attrs: AttributeSet) -> Arc<dyn BoundedMeasure<T>>;
+}
+
+impl<F, T> BoundedMeasureGenerator<T> for F
+where
+    F: Fn(AttributeSet) -> Arc<dyn BoundedMeasure<T>> + Send + Sync + 'static,
+{
+    fn generate(&self, attrs: AttributeSet) -> Arc<dyn BoundedMeasure<T>> {
+        self(attrs)
+    }
+}
+
+/// Receives measurements to be aggregated for a predetermined attribute set
+pub(crate) trait BoundedMeasure<T>: Send + Sync + 'static {
+    fn call(&self, measurement: T);
+}
+
+impl<F, T> BoundedMeasure<T> for F
+where
+    F: Fn(T) + Send + Sync + 'static,
+{
+    fn call(&self, measurement: T) {
+        self(measurement)
     }
 }
 
@@ -105,11 +137,18 @@ impl<T: Number<T>> AggregateBuilder<T> {
     /// Builds a last-value aggregate function input and output.
     ///
     /// [Builder::temporality] is ignored and delta is always used.
-    pub(crate) fn last_value(&self) -> (impl Measure<T>, impl ComputeAggregation) {
+    pub(crate) fn last_value(
+        &self,
+    ) -> (
+        impl Measure<T>,
+        impl ComputeAggregation,
+        impl BoundedMeasureGenerator<T>,
+    ) {
         // Delta temporality is the only temporality that makes semantic sense for
         // a last-value aggregate.
         let lv_filter = Arc::new(LastValue::new());
         let lv_agg = Arc::clone(&lv_filter);
+        let lv_bound = Arc::clone(&lv_filter);
 
         (
             self.filter(move |n, a| lv_filter.measure(n, a)),
@@ -128,6 +167,7 @@ impl<T: Number<T>> AggregateBuilder<T> {
 
                 (g.data_points.len(), new_agg.map(|a| Box::new(a) as Box<_>))
             },
+            move |attrs: AttributeSet| last_value::generate_bound_measure(&lv_bound, attrs),
         )
     }
 
@@ -135,9 +175,14 @@ impl<T: Number<T>> AggregateBuilder<T> {
     pub(crate) fn precomputed_sum(
         &self,
         monotonic: bool,
-    ) -> (impl Measure<T>, impl ComputeAggregation) {
+    ) -> (
+        impl Measure<T>,
+        impl ComputeAggregation,
+        impl BoundedMeasureGenerator<T>,
+    ) {
         let s = Arc::new(PrecomputedSum::new(monotonic));
         let agg_sum = Arc::clone(&s);
+        let bound_sum = Arc::clone(&s);
         let t = self.temporality;
 
         (
@@ -145,14 +190,25 @@ impl<T: Number<T>> AggregateBuilder<T> {
             move |dest: Option<&mut dyn Aggregation>| match t {
                 Some(Temporality::Delta) => agg_sum.delta(dest),
                 _ => agg_sum.cumulative(dest),
+            },
+            move |attrs: AttributeSet| {
+                sum::generate_bound_measure_precomputed_sum(&bound_sum, attrs)
             },
         )
     }
 
     /// Builds a sum aggregate function input and output.
-    pub(crate) fn sum(&self, monotonic: bool) -> (impl Measure<T>, impl ComputeAggregation) {
+    pub(crate) fn sum(
+        &self,
+        monotonic: bool,
+    ) -> (
+        impl Measure<T>,
+        impl ComputeAggregation,
+        impl BoundedMeasureGenerator<T>,
+    ) {
         let s = Arc::new(Sum::new(monotonic));
         let agg_sum = Arc::clone(&s);
+        let bound_sum = Arc::clone(&s);
         let t = self.temporality;
 
         (
@@ -161,6 +217,7 @@ impl<T: Number<T>> AggregateBuilder<T> {
                 Some(Temporality::Delta) => agg_sum.delta(dest),
                 _ => agg_sum.cumulative(dest),
             },
+            move |a| sum::generate_bound_measure_sum(&bound_sum, a),
         )
     }
 
@@ -170,9 +227,14 @@ impl<T: Number<T>> AggregateBuilder<T> {
         boundaries: Vec<f64>,
         record_min_max: bool,
         record_sum: bool,
-    ) -> (impl Measure<T>, impl ComputeAggregation) {
+    ) -> (
+        impl Measure<T>,
+        impl ComputeAggregation,
+        impl BoundedMeasureGenerator<T>,
+    ) {
         let h = Arc::new(Histogram::new(boundaries, record_min_max, record_sum));
         let agg_h = Arc::clone(&h);
+        let bound_h = Arc::clone(&h);
         let t = self.temporality;
 
         (
@@ -181,6 +243,7 @@ impl<T: Number<T>> AggregateBuilder<T> {
                 Some(Temporality::Delta) => agg_h.delta(dest),
                 _ => agg_h.cumulative(dest),
             },
+            move |attrs: AttributeSet| histogram::generate_bound_measure(&bound_h, attrs),
         )
     }
 
@@ -191,7 +254,11 @@ impl<T: Number<T>> AggregateBuilder<T> {
         max_scale: i8,
         record_min_max: bool,
         record_sum: bool,
-    ) -> (impl Measure<T>, impl ComputeAggregation) {
+    ) -> (
+        impl Measure<T>,
+        impl ComputeAggregation,
+        impl BoundedMeasureGenerator<T>,
+    ) {
         let h = Arc::new(ExpoHistogram::new(
             max_size,
             max_scale,
@@ -199,6 +266,7 @@ impl<T: Number<T>> AggregateBuilder<T> {
             record_sum,
         ));
         let agg_h = Arc::clone(&h);
+        let bound_h = Arc::clone(&h);
         let t = self.temporality;
 
         (
@@ -206,6 +274,9 @@ impl<T: Number<T>> AggregateBuilder<T> {
             move |dest: Option<&mut dyn Aggregation>| match t {
                 Some(Temporality::Delta) => agg_h.delta(dest),
                 _ => agg_h.cumulative(dest),
+            },
+            move |attrs: AttributeSet| {
+                exponential_histogram::generate_bound_measure(&bound_h, attrs)
             },
         )
     }