--- a/opentelemetry-sdk/src/metrics/internal/mod.rs
+++ b/opentelemetry-sdk/src/metrics/internal/mod.rs
@@ -8,7 +8,8 @@ use core::fmt;
 use std::ops::{Add, AddAssign, Sub};
 
 pub(crate) use aggregate::{
-    AggregateBuilder, BoundedMeasure, BoundedMeasureGenerator, ComputeAggregation, Measure, MeasureSet
+    AggregateBuilder, BoundedMeasure, BoundedMeasureGenerator, ComputeAggregation, Measure,
+    MeasureSet,
 };
 pub(crate) use exponential_histogram::{EXPO_MAX_SCALE, EXPO_MIN_SCALE};