… opentelemetry-sdk/src/metrics/mod.rs
… opentelemetry-sdk/src/metrics/pipeline.rs
… opentelemetry/src/metrics/instruments/counter.rs
… opentelemetry/src/metrics/instruments/up_down_counter.rs
… opentelemetry/src/metrics/noop.rs
… stress/Cargo.toml
… stress/src/metrics_bounded.rs