Fixes #
Design discussion issue (if applicable) #1374

## Changes

This PR adds API support for bound instruments for `Counter` and `UpDownCounter`.  Bound instruments are counters which always change the value for a specific and predefined attribute set.  The focus on this change is creating the initial API surface for bound instruments, and the mechanisms to allow non-bound instruments to become bound (since there is no direct path from aggregates to counters, it's all through closures).  

To facilitate counters creating bound instruments that can refer back to their source aggregates, aggregates return a `BoundMeasureGenerator<T>` closure that is provided alongside the `Measure<T>` for the aggregate.  This closure can then be invoked in order to return a `BoundMeasure<T>`, which provides the mechanism to invoke a measure for the cached attribute set. 

This PR is *not* performance focused.  It should not cause regressions but using bound instruments from this PR may not cause increased throughput.  Performance enhancements that take advantage of bound instruments will come in a follow-up PR once we have an agreed upon API and generation path.  

## Benchmarks/Stress Tests

### Main

```
Counter_Add_Sorted      time:   [717.94 ns 723.06 ns 729.17 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

Counter_Add_Unsorted    time:   [711.21 ns 714.89 ns 719.15 ns]
Found 12 outliers among 100 measurements (12.00%)
  8 (8.00%) high mild
  4 (4.00%) high severe
```

```
stress test: metrics
Number of threads: 6
Throughput: 6,367,800 iterations/sec
Throughput: 6,581,800 iterations/sec
Throughput: 6,736,800 iterations/sec
Throughput: 6,720,200 iterations/sec
Throughput: 6,804,600 iterations/sec
```

### PR

```
Counter_Add_Sorted      time:   [712.98 ns 715.33 ns 717.99 ns]
                        change: [-0.7779% +0.3569% +1.5923%] (p = 0.60 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe

Counter_Add_Unsorted    time:   [705.13 ns 707.71 ns 710.82 ns]
                        change: [-2.0962% -0.9670% +0.0233%] (p = 0.09 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe

Counter_Bounded         time:   [133.96 ns 134.82 ns 135.86 ns]
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) high mild
  9 (9.00%) high severe
```

```
stress test: metrics
Number of threads: 6
Throughput: 6,173,200 iterations/sec
Throughput: 6,381,600 iterations/sec
Throughput: 6,250,000 iterations/sec
Throughput: 6,434,000 iterations/sec
```

```
stress test: metrics_bounded
Number of threads: 6
Throughput: 6,473,400 iterations/sec
Throughput: 6,459,600 iterations/sec
Throughput: 6,339,200 iterations/sec
Throughput: 6,369,200 iterations/sec
```

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [X] Unit tests added/updated (if applicable)
* [ ] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [ ] Changes in public API reviewed (if applicable)