@KallDrexx 2023-12-14
The api will change slightly when/if #1421 get in to take an attribute set instead of a slice of key value pairs, but I  didn't want to couple both together.  Bounded performance might end up getting a bit better since `cloning()` becomes cheaper, but performance isn't the point of this PR :)