@KallDrexx 2023-12-19
> Like the general direction, kinda surprised not to see much of a performance gain over standard counters. I feel like (not related to this change) we have a lot of Arc spread through the code that maybe we could avoid. A bit of locking our locks so we can lock more locks.

`Arc` contains no locking.  It's an atomic number that tracks how many references there are.  When you `clone()` it increases the atomic count, and when you `drop()` it decrements the atomic count (and if it's the last reference it drops the underlying value).  There's no locking involved in cloning, dropping, or accessing an `Arc`, and a lock is only involved if you access an internal function of the held type that invokes a lock.  

The reason this doesn't provide any performance benefits is partially because #1421 is not yet merged.  Since we don't have that change in, we must do a hard clone of the attribute set every time a bound measure is called (and for each measure in the `ResolvedMeasure`.  Without #1421 it's cloning the whole vector and doing extra allocations.    

There's no performance benefit too because this PR has bound instruments as a thin wrapper around normal measurements.  Once this PR is merged, the next PR fully takes advantage of bound instruments by creating an atomic that's incremented by the bound measure function.  This allows the `BoundCounter.add()` calls to increment with zero mutex locking and zero hashmap lookups.  That's what this PR is setting the stage for.