@cijothomas 2024-07-02
> @cijothomas Can we close this PR if you plan to open a new PR for bound instruments inline with the changes here?

Yes closing this. Cross-linking to https://github.com/open-telemetry/opentelemetry-rust/issues/1374, so its is easy to find the work in this PR to adopt.